Page-locked memory
Purpose of page-locked memory:
Page-locked memory/pinned memory, is a type of memory allocation in CUDA that cannot be moved by the OS
it is kept pinned in physicl RAM. Allowing for the GPU to directly access and read/write the data without
copying it to a buffer first. This significantly speeds up the transfer of data between the CPU and GPU. One downside
is that page-locked memory takes up system memory, and if used excesively can go over amount of memory available.